+++ bika/lims/browser/contentmenu.py
@@ -21,7 +21,7 @@
 from plone.app.contentmenu.menu import WorkflowMenu as BaseClass
 from plone.app.contentmenu.view import ContentMenuProvider
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
-from bika.lims.interfaces import IHaveNoActionsMenu
+from bika.lims.interfaces import IHideActionsMenu
 from zope.component import getUtility
 from zope.browsermenu.interfaces import IBrowserMenu
 
@@ -38,7 +38,7 @@ class SenaiteContentMenuProvider(ContentMenuProvider):
     # From IContentMenuView
 
     def available(self):
-        if IHaveNoActionsMenu.providedBy(self.context):
+        if IHideActionsMenu.providedBy(self.context):
             return False
         return True